section 1.3: created HomeViewModel and edited HomeView
