+++ activeMeApp/Home/Models/Workout.swift
@@ -1,0 +1,19 @@
+//
+//  Workout.swift
+//  activeMeApp
+//
+//  Created by Tahera Akter Mukta on 02/02/2025.
+//
+
+import SwiftUI
+
+struct Workout {
+    let id: Int
+    let tital: String
+    let image: String
+    let tintColor: Color
+    let duration: String
+    let date: String
+    let calories: String
+}
+
