+++ activeMeApp/Home/Views/WorkoutCardView.swift
@@ -7,15 +7,6 @@
 
 import SwiftUI
 
-struct Workout {
-    let id: Int
-    let tital: String
-    let image: String
-    let tintColor: Color
-    let duration: String
-    let date: String
-    let calories: String
-}
 
 struct WorkoutCardView: View {
     @State var workout: Workout
